requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.0
pyjwt==2.8.0
bcrypt==4.1.3
stripe==9.9.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
cryptography==42.0.5
